fix: airbag methods. modularize code.
