+++ src/vin/mod.rs
@@ -1,2 +1,6 @@
+pub mod attributes;
 pub mod element_ids;
 pub mod parser;
+pub mod pattern;
+pub mod schema;
+pub mod wmi;
